Remove hard coding of osuosl.org domain for postfix
This allows for hosts that are not osuosl.org (i.e. phpbb.com) have their mail
come from the correct domain. By default the postfix cookbook sets the
``mydomain`` setting using the ohai fact. For now lets override the ``myorigin``
setting not to be the FQDN but instead be the ``$mydomain`` setting.

For now keep the relayhost set directly to be ``osuosl.org``. That ensures we
don't have to muck with DNS on non-standard subdomains.

This should correctly resolve osuosl-cookbooks/proj-phpbb#62 finally.