Reordered search domains in workstation role.
After much research, I found the workstation role had settings for
resolver.  I do not know how to test this change -- it would be nice
if I could spin up a virtual workstation or something and check this
change -- but the existing settings exactly match what is in the
resolv.conf file on neon so the new settings should work.