Create a base workstation role for staff
This creates a workstation_base role that can be used by staff to run chef on
their workstations setup in a similar fashion as the student workstations. The
only difference is removing the nfs, ldap and other related recipes.

This also changes the workstation role to just include the workstation_base role
and keep the recipes it needs.