Move apt/repo recipes to before almost everything else
I've noticed during a boostrap of a new workstation I get the following error:

WARNING: The following packages cannot be authenticated!

This seems to be stem from a problem with not updating the apt cache prior to
running apt-get install. To work around this, we should ensure we get our repos
in place first which will force an update anyways. Probably makes sense to add
all of these repos first anyways.