berks-env-check script to check that all environments resolve properly
This is a script that can be run locally or by jenkins eventually at the root of
the chef-repo. It runs ``berks install`` on each enviroment to ensure all of the
depenendancies are resolved. This points to our chef server so if there are any
problems in production this should trigger that.

This also includes various fixes to other environments which currently fail this
script.