Update various upstream cookbooks which fix Chef 13 deprecations
These are the cookbooks that need to be updated in order to get the base_managed
role to pass Chef 13 deprecations properly. Some of these are dependencies that
are required for some of the updated cookbooks.

For some of these environments, we needed to stick with older versions due to
their own requirements. I *believe* these are safe changes from what I can tell.